requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
bitarray==2.9.2
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
ckzg==2.0.0
click==8.1.7
colorlog==6.8.2
constantly==23.10.4
cryptography==43.0.0
cytoolz==0.12.3
daphne==4.1.2
dataclasses-json==0.6.7
distro==1.9.0
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
eth-account==0.13.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0
fastapi==0.115.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
langsmith==0.1.141
marshmallow==3.23.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.54.3
orjson==3.10.11
packaging==24.1
parsimonious==0.10.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-settings==2.6.1
pydantic_core==2.20.1
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
regex==2024.7.24
requests==2.31.0
requests-toolbelt==1.0.0
rlp==4.0.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
starlette==0.38.6
tenacity==9.0.0
tiktoken==0.8.0
tomli==2.0.1
toolz==0.12.1
tqdm==4.67.0
Twisted==24.7.0
txaio==23.1.1
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.31.0
web3==7.0.0
websockets==13.0
yarl==1.9.4
zope.interface==7.0.1
